Add files via upload
to use, unzip the file open it in VS code and change the SSID and password to that of your local network. connect your ESP 32 with humidity and temp sensor type in the terminal node server.js to start the server, and paste the link it gives you in there terminal into your browser.